roof/requirements.txt
# Roof Detection Service Dependencies
# �ݶ�����������

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML
ultralytics==8.0.200
torch>=2.0.1,<3.0.0
torchvision>=0.15.2,<1.0.0

# Image processing
opencv-python-headless==4.8.1.78
Pillow==10.0.0
numpy==1.24.3

# Geometry
shapely==2.0.1
